backend/pyproject.toml
[project]
name = "psa-backend"
version = "0.1.0"
description = "PSA Agent Backend with Strands Multi-Agent Architecture"
authors = [
    {name = "Vista Hackathon Team"}
]
readme = "README.md"
requires-python = "^3.11"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "strands-agents (>=0.1.0)",
    "strands-tools (>=0.1.0)",
    "python-dotenv (>=1.0.0)",
    "pydantic (>=2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
